refactor: add file validation to only images and videos
