+++ src/components/InputMedia/InputMedia.module.scss
@@ -22,7 +22,7 @@ img {
   height: auto;
 }
 
-.hiden {
+.hidden {
   display: none;
 }
